+++ app/Http/Middleware/Authenticate.php
@@ -14,8 +14,11 @@ class Authenticate extends Middleware
      */
     protected function redirectTo($request)
     {
-        if (! $request->expectsJson()) {
-            return route('login');
+        if (!$request->wantsJson()) {
+            return redirect()->guest(route('auth/login'));
         }
+
+        return response()->json(['message' => 'Invalid API token'], 401);
+
     }
 }
