--- database/seeds/UserSeeder.php
@@ -1,20 +1,0 @@
-<?php
-
-use Illuminate\Database\Seeder;
-
-class UserSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        DB::table('users')->insert([
-            'username' => "admin",
-            'email' => "admin-user@example.org",
-            'password' => Hash::make('admin'),
-        ]);
-    }
-}
