+++ resources/js/components/upload/UploadManager.vue
@@ -1,8 +1,6 @@
 <template></template>
 
 <script>
-    import UploadService from "@js/services/UploadFilesService";
-
     export default {
         data(){
             return {
@@ -68,6 +66,8 @@
                 //Grab the first file from the stack
                 var currentFile = this.files.shift();
 
+                //TODO check if file is a valid media format
+
                 //Update status variables
                 this.updateStatusVariables();
 
@@ -75,11 +75,20 @@
                 this.isUploading = true;
                 this.status.currentFile = currentFile.name;
 
+                let formData = new FormData();
+                formData.append("media", currentFile);
+
                 //Upload the file
-                UploadService.upload(currentFile, event => {
-                    this.status.progress = Math.round((100 * event.loaded) / event.total);
+                axios.post("/api/upload", formData, {
+                    headers: {
+                        "Content-Type": "multipart/form-data"
+                    },
+                    onUploadProgress: () => {
+                        this.status.progress = Math.round((100 * event.loaded) / event.total);
+                    }
                 })
                 .then(response => {
+                    console.log("upload response: ", response);
                     //On success?
                     //this.message = response.data.message;
                 })
