--- resources/js/services/UploadFilesService.js
@@ -1,21 +1,0 @@
-/**
- * Credits: https://github.com/bezkoder/vue-upload-files
- */
-import http from "@js/http-common";
-
-class UploadFilesService {
-    upload(file, onUploadProgress) {
-        let formData = new FormData();
-
-        formData.append("media", file);
-
-        return http.post("/upload", formData, {
-            headers: {
-                "Content-Type": "multipart/form-data"
-            },
-            onUploadProgress
-        });
-    }
-}
-
-export default new UploadFilesService();
