+++ routes/api.php
@@ -14,17 +14,16 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+
 Route::post('auth/login', 'AuthApiController@login');
 Route::post('auth/register', 'AuthApiController@register');
 
-Route::group(['middleware' => 'auth'], function () {
-    Route::group(['middleware' => 'auth:api'], function () {
-        Route::get('auth/logout', 'AuthApiController@logout');
-        Route::get('auth/user', 'AuthApiController@user');
-    });
+Route::group(['middleware' => 'auth:api'], function () {
+    Route::get('auth/logout', 'AuthApiController@logout');
+    Route::get('auth/user', 'AuthApiController@user');
+    Route::post('upload', 'UploadApiController@store');
 });
 
-Route::post('upload', 'UploadApiController@store');
 Route::get('album', 'AlbumApiController@get');
 Route::get('artist', 'ArtistApiController@get');
 Route::get('request', 'RequestApiController@index');
